chore(tests): move tests to src folder
