+++ src/test-utils/jestGlobalSetup.ts
@@ -1,0 +1,3 @@
+import * as matchers from "./matchers";
+
+expect.extend(matchers);
