--- tests/jestGlobalSetup.ts
@@ -1,3 +1,0 @@
-import * as matchers from "./utils/matchers";
-
-expect.extend(matchers);
