--- tests/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "../",
-    "types": ["@types/jest", "node"]
-  },
-  "include": ["./"],
-}
